blitz_cli/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   groq
    #   httpx
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
colorama==0.4.6
    # via loguru
distro==1.9.0
    # via groq
filetype==1.2.0
    # via langchain-google-genai
google-ai-generativelanguage==0.6.18
    # via langchain-google-genai
google-api-core==2.25.1
    # via google-ai-generativelanguage
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.3
    # via sqlalchemy
groq==0.28.0
    # via langchain-groq
grpcio==1.73.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.73.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   langgraph-sdk
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.26
    # via
    #   my-project (pyproject.toml)
    #   langchain-sambanova
langchain-core==0.3.66
    # via
    #   langchain
    #   langchain-google-genai
    #   langchain-groq
    #   langchain-sambanova
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==2.1.5
    # via my-project (pyproject.toml)
langchain-groq==0.3.4
    # via my-project (pyproject.toml)
langchain-sambanova==0.1.5
    # via my-project (pyproject.toml)
langchain-text-splitters==0.3.8
    # via langchain
langgraph==0.4.9
    # via my-project (pyproject.toml)
langgraph-checkpoint==2.1.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.2.2
    # via langgraph
langgraph-sdk==0.1.70
    # via langgraph
langsmith==0.4.1
    # via
    #   langchain
    #   langchain-core
loguru==0.7.3
    # via my-project (pyproject.toml)
orjson==3.10.18
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.31.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   groq
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via my-project (pyproject.toml)
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
requests==2.32.4
    # via
    #   google-api-core
    #   langchain
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   groq
sqlalchemy==2.0.41
    # via langchain
sseclient-py==1.8.0
    # via langchain-sambanova
tenacity==9.1.2
    # via langchain-core
types-requests==2.32.4.20250611
    # via langchain-sambanova
typing-extensions==4.14.0
    # via
    #   anyio
    #   groq
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
win32-setctime==1.2.0
    # via loguru
xxhash==3.5.0
    # via langgraph
zstandard==0.23.0
    # via langsmith
